// src/app/helpers/model/policePraemienzahler.ts
/**
 * CSS customer portal API: Vertrag
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { PoliceProdukt } from './policeProdukt';


export interface PolicePraemienzahler { 
    /**
     * product in policy
     */
    policeProdukt: Array<PoliceProdukt>;
    /**
     * price of this product
     */
    subtotal: number;
    pzvornameNachname?: string;
}