// src/app/helpers/model/policeProdukt.ts
/**
 * CSS customer portal API: Vertrag
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface PoliceProdukt { 
    /**
     * version of terms and conditions
     */
    avbVersion?: string;
    /**
     * premium in document
     */
    dokumentenPraemie?: number;
    gesetzLabel?: string;
    /**
     * containing law (KVG / VVG)
     */
    gesetz: PoliceProdukt.GesetzEnum;
    /**
     * product name as string in users portal language
     */
    produktNameLabel: string;
    /**
     * additional information when the product was released
     */
    zbVersion?: string;
}
export namespace PoliceProdukt {
    export type GesetzEnum = 'IVG' | 'KVG' | 'MVG' | 'OR' | 'UNBEKANNT' | 'UVG' | 'VVG';
    export const GesetzEnum = {
        IVG: 'IVG' as GesetzEnum,
        KVG: 'KVG' as GesetzEnum,
        MVG: 'MVG' as GesetzEnum,
        OR: 'OR' as GesetzEnum,
        UNBEKANNT: 'UNBEKANNT' as GesetzEnum,
        UVG: 'UVG' as GesetzEnum,
        VVG: 'VVG' as GesetzEnum
    };
}