// src/app/helpers/model/spvPolicy.ts
/**
 * CSS customer portal API: Vertrag
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { SpvProdukt } from './spvProdukt';


/**
 * A Spv policy
 */
export interface SpvPolicy { 
    /**
     * Spv Policy unique identifier
     */
    policenNr: string;
    /**
     * document identifier to load from resouce /dokumentenverwaltung
     */
    dokumentNr?: string;
    /**
     * true if shv police document is available
     */
    dokumentAvailable: boolean;
    /**
     * date as String: yyyy-MM-dd
     */
    vertragsbeginn: string;
    /**
     * date as String: yyyy-MM-dd
     */
    vertragsende: string;
    /**
     * true if spv police will be extended automatically
     */
    automatischeVerlaengerung: boolean;
    /**
     * zahlungsperiode
     */
    zahlungsPeriodeLabel: string;
    /**
     * Erstellungsgrund
     */
    erstellungsgrundLabel: string;
    produkt: SpvProdukt;
    /**
     * how is the police paid 
     */
    zahlwegLabel?: string;
    /**
     * PartnerNr as label (with dashes)
     */
    partnerNrLabel: string;
}